+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bluetooth: New methods `EspBleGap::start_scanning` and `EspBleGap::stop_scanning`
 - New example, `bt_ble_gap_scanner` to demonstrate usage of added ble scanning methods
 - New example, `mdns_advertise` to demonstrate mDNS service advertisement
+- Wifi: New methods `set_max_tx_power`, `get_max_tx_power`
 
 ## [0.51.0] - 2025-01-15